Hadoop第一个demo跑通了

+++ src/main/java/com/matcha/test/mapper/OrderIntTestMapper.java
@@ -1,6 +1,7 @@
 package com.matcha.test.mapper;
 
 import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 
@@ -9,10 +10,10 @@ import java.io.IOException;
 /**
  * Created by Administrator on 2016/10/24.
  */
-public class OrderIntTestMapper extends Mapper<IntWritable, Text, IntWritable, IntWritable>
+public class OrderIntTestMapper extends Mapper<LongWritable, Text, IntWritable, IntWritable>
 {
     @Override
-    protected void map(IntWritable key, Text value, Context context) throws IOException, InterruptedException
+    protected void map(LongWritable key, Text value, Context context) throws IOException, InterruptedException
     {
         System.out.println("read data key - " + key.get() + " value - " + value.toString());
         String valueStr = value.toString();
